🐛 [BE] dev 빌드 충돌 해소(admin/common 복원) + 자진삭제 글 벌점 부과
- admin/common/AdminServiceSupport 복원(+escapeLike): dev 의 AdminDashboardServiceImpl(PR #81)이
  이 클래스를 쓰고 있어, 도메인별 3벌 분리 상태로 머지하면 컴파일이 깨진다.
  PostAdminSupport/CommentAdminSupport/SanctionAdminSupport 삭제하고 호출부 3곳을 공통 헬퍼로 되돌림.
  (분리 사유였던 "MyBatis type-alias 충돌"은 단일 클래스명이라 성립하지 않음)
- softDelete: "이미 deleted" 를 살아있는 벌점 유무로 판별. 작성자가 신고를 눈치채고 먼저 지운 글에도
  조치 이력·벌점을 남긴다. 관리자가 이미 삭제 조치한 대상만 no-op → 벌점 중복 부과는 그대로 차단.
- CLAUDE.md 패키지 구조 갱신 (admin/comment 추가, common·sanction 소속 정리)
