+++ src/main/java/com/back/admin/common/AdminServiceSupport.java
@@ -1,16 +1,15 @@
-package com.back.admin.post.support;
+package com.back.admin.common;
 
 import com.back.global.exception.BaseException;
 import com.back.global.security.AuthUtils;
 
-// 게시글 관리 서비스 공통 헬퍼 (인증 주체 추출, 일괄 실패 메시지, 페이지 파라미터 보정).
-// 상태를 갖지 않는 정적 유틸. 이름은 도메인별로 유일해야 한다 —
-// mybatis.type-aliases-package=com.back 가 단순 클래스명으로 별칭을 등록하므로 중복명은 기동 실패를 유발한다.
-public final class PostAdminSupport {
+// 관리자 서비스 공통 헬퍼 (인증 주체 추출, 일괄 실패 메시지, 페이지 파라미터 보정, LIKE 이스케이프).
+// 상태를 갖지 않는 정적 유틸. 관리자 화면(게시글·댓글·신고/제재·대시보드)이 공유한다.
+public final class AdminServiceSupport {
 
     private static final int MAX_PAGE_SIZE = 100;
 
-    private PostAdminSupport() {
+    private AdminServiceSupport() {
     }
 
     // 현재 로그인한 관리자 user_id (/api/admin/** 는 SecurityConfig 에서 ADMIN 으로 이미 제한됨)
@@ -34,5 +33,16 @@ public final class PostAdminSupport {
             return 1;
         }
         return Math.min(size, MAX_PAGE_SIZE);
+    }
+
+    // LIKE 검색어 이스케이프: 사용자가 입력한 % _ \ 를 리터럴로 취급하도록 백슬래시 이스케이프.
+    // 매퍼의 LIKE ... ESCAPE '\\' 와 짝을 이룬다. null/공백이면 null 반환(검색 미적용).
+    public static String escapeLike(String keyword) {
+        if (keyword == null || keyword.isBlank()) {
+            return null;
+        }
+        return keyword.replace("\\", "\\\\")
+                      .replace("%", "\\%")
+                      .replace("_", "\\_");
     }
 }
